refactor: Update CustomerForm.php, TransactionTable.php, CreateCustomer.php, PlanIndex.php, SettingComponent.php, and CustomerTable.php
Refactor multiple files to improve code readability and functionality. Changes include:
- CustomerForm.php: Updated validation rules for 'isolir_date' field.
- TransactionTable.php: Updated column label from 'Tagihan' to 'Tenggat' and formatted 'due_date' to display day, month, and year.
- CreateCustomer.php: Updated logic for setting 'isolirDate' based on 'isolir_date' value.
- PlanIndex.php: Added exception handling for deleting plans.
- SettingComponent.php: Updated logic for storing 'ppn' value and updating 'tax_rate' and 'total_amount' for unpaid bills.
- CustomerTable.php: Added sorting functionality for 'Name' column and formatted 'isolir_date' to display 'Akhir Bulan' for 'last_day' value.

Fixes #123

+++ app/Models/Profile.php
@@ -8,6 +8,7 @@ use Illuminate\Database\Eloquent\Model;
 use RouterOS\Client;
 use RouterOS\Query;
 use Sushi\Sushi;
+use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class Profile extends Model
 {
@@ -115,10 +116,15 @@ class Profile extends Model
                 ->equal('.id', $id);
             $response = $client->query($query)->read();
             if (\is_array($response) && !empty($response)) {
+                if (\str_contains($response['after']['message'], 'no such item')) {
+                    throw new NotFoundResourceException();
+                }
                 throw new Exception('Failed to delete from Mikrotik');
             }
 
             return !isset($response['after']['ret']);
+        } catch (NotFoundResourceException $th) {
+            throw $th;
         } catch (\Throwable $th) {
             \info('error', ['error' => $th->getMessage()]);
             return false;
